+++ packages/stryker-babel-transpiler/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.1"></a>
+## [0.3.1](https://github.com/stryker-mutator/stryker/compare/stryker-babel-transpiler@0.3.0...stryker-babel-transpiler@0.3.1) (2018-02-07)
+
+
+### Bug Fixes
+
+* **babel-transpiler:** log babel config on debug ([#614](https://github.com/stryker-mutator/stryker/issues/614)) ([393db1f](https://github.com/stryker-mutator/stryker/commit/393db1f))
+* **dependencies:** update stryker-api requirement to ^0.13.0 ([8eba6d4](https://github.com/stryker-mutator/stryker/commit/8eba6d4))
+
+
+
+
+<a name="0.3.0"></a>
+# [0.3.0](https://github.com/stryker-mutator/stryker/compare/stryker-babel-transpiler@0.2.4...stryker-babel-transpiler@0.3.0) (2018-02-07)
+
+
+### Features
+
+* **coverage analysis:** Support transpiled code ([#559](https://github.com/stryker-mutator/stryker/issues/559)) ([7c351ad](https://github.com/stryker-mutator/stryker/commit/7c351ad))
+
+
+
+
 <a name="0.2.4"></a>
 ## [0.2.4](https://github.com/stryker-mutator/stryker/compare/stryker-babel-transpiler@0.2.3...stryker-babel-transpiler@0.2.4) (2018-01-15)