+++ packages/stryker-html-reporter/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.1"></a>
+## [0.12.1](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.12.0...stryker-html-reporter@0.12.1) (2018-02-07)
+
+
+### Bug Fixes
+
+* **dependencies:** update stryker-api requirement to ^0.13.0 ([8eba6d4](https://github.com/stryker-mutator/stryker/commit/8eba6d4))
+
+
+
+
+<a name="0.12.0"></a>
+# [0.12.0](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.11.5...stryker-html-reporter@0.12.0) (2018-02-07)
+
+
+### Features
+
+* **coverage analysis:** Support transpiled code ([#559](https://github.com/stryker-mutator/stryker/issues/559)) ([7c351ad](https://github.com/stryker-mutator/stryker/commit/7c351ad))
+
+
+
+
 <a name="0.11.5"></a>
 ## [0.11.5](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.11.4...stryker-html-reporter@0.11.5) (2018-01-24)