+++ packages/stryker-karma-runner/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.2"></a>
+## [0.12.2](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.12.1...stryker-karma-runner@0.12.2) (2018-02-07)
+
+
+### Bug Fixes
+
+* **dependencies:** update stryker-api requirement to ^0.13.0 ([8eba6d4](https://github.io/stryker-mutator/stryker/commit/8eba6d4))
+
+
+
+
+<a name="0.12.1"></a>
+## [0.12.1](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.12.0...stryker-karma-runner@0.12.1) (2018-02-07)
+
+
+
+
+**Note:** Version bump only for package stryker-karma-runner
+
 <a name="0.12.0"></a>
 # [0.12.0](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.11.3...stryker-karma-runner@0.12.0) (2018-01-10)